debuged

+++ src/main/java/anbar/model/service/ProductService.java
@@ -14,21 +14,21 @@ public class ProductService {
 
     public static void edit(Product product) throws Exception {
         try(ProductRepository productRepository = new ProductRepository()){
-//            if (productRepository.findById(product.getId()) != null) {
+            if (productRepository.findById(product.getId()) != null) {
                 productRepository.edit(product);
-//            }else{
-//                throw new Exception("Product not found");   // ProductNotFoundException
-//            }
+            }else{
+                throw new Exception("Product not found");
+            }
         }
     }
 
     public static void delete(int id) throws Exception {
         try(ProductRepository productRepository = new ProductRepository()){
-//            if (productRepository.findById(product.getId()) != null) {
+            if (productRepository.findById(id) != null) {
                 productRepository.delete(id);
-//            }else{
-//                throw new Exception("Product not found");   // ProductNotFoundException
-//            }
+            }else{
+                throw new Exception("Product not found");
+            }
         }
     }
 
@@ -38,7 +38,17 @@ public class ProductService {
         }
     }
 
-//    todo : findByTitle Like
+    public static Product findById(int id) throws Exception {
+        try(ProductRepository ProductRepository = new ProductRepository()){
+            return ProductRepository.findById(id);
+        }
+    }
+
+public static List<Product> findByTitle(String title) throws Exception {
+    try(ProductRepository productRepository = new ProductRepository()){
+        return productRepository.findByTitle(title);
+    }
+}
 
     public static List<Product> findByPrice(int price1,int price2) throws Exception {
         try(ProductRepository productRepository = new ProductRepository()){
